Fix invalid constructor call for ArcanistHardpointFutureList
Summary:
`ArcanistHardpointFutureList` does not have a constructor and must be instantiated without any parameters.
As the code checks `if ($result instanceof Future)` and tries to pass `$result` as a parameter, the intention seems to be calling `newFromFutures($result)` on the new `ArcanistHardpointFutureList` instance.

Closes T15836

Test Plan: Read the code in `ArcanistHardpointFutureList`.

Reviewers: O1 Blessed Committers, valerio.bozzolan

Reviewed By: O1 Blessed Committers, valerio.bozzolan

Subscribers: speck, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T15836

Differential Revision: https://we.phorge.it/D25708