xhpast: Accept CXXFLAGS and CPPFLAGS from the environment
Summary:
This allows users to pass customized flags to the compiler.

The motivation for this change is to include flags provided by automatic packaging systems.

For example, the build system of AOSC OS sets CXXFLAGS to tell the compiler to do cross-compile (when necessary) and generate debuginfo.

Test Plan: Provide some cross-compilation flags in CXXFLAGS and observe the binary.

Reviewers: O1 Blessed Committers, valerio.bozzolan

Reviewed By: O1 Blessed Committers, valerio.bozzolan

Subscribers: aklapper, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Tags: #arcanist

Differential Revision: https://we.phorge.it/D26070