Czech Locale: Do not crash on missing plural translations
Summary:
When a string with plural forms has not yet been translated to Czech language (and thus falls back to English language with one singular and one plural), do not crash on accessing the missing second plural form in the array. Instead, just reuse the first plural form for the second plural form (which is the same in English language).

Closes T16172

Test Plan:
* Install translations.
* Go to https://phabricator.wikimedia.org/settings/panel/language/ and set your language to Czech (which has more than one plural form).
* Go to a page which displays a string with plural forms and that string has not yet been translated to Czech. I guess.

Reviewers: O1 Blessed Committers, valerio.bozzolan

Reviewed By: O1 Blessed Committers, valerio.bozzolan

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16172

Differential Revision: https://we.phorge.it/D26182