Diviner: Update "Developing XHPAST" article
Summary:
Maybe flex version 2.3.35 is required (if that is not a typo), however version 2.5.35 was already released February 26, 2008.
You probably don't want to develop XHPAST on a system which is more than 17 years old, so this is irrelevant nowadays.
Also flex moved away from Sourceforge in the meantime.

Test Plan:
* View https://flex.sourceforge.net/ and https://en.wikipedia.org/w/index.php?title=Flex_(lexical_analyser_generator)&oldid=398881651
* Optionally run `./bin/diviner generate` in Phorge and go to http://phorge.localhost/book/arcanist/article/developing_xhpast/

Reviewers: O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Differential Revision: https://we.phorge.it/D26258