arc weld: clarify it's a joke workflow
Summary:
Surprise surprise, the 'arc weld' workflow was yet another nonsense demo
doing nothing useful. So, let's harden the joke to be more self-explaining.

New output:

    $ arc weld --help
    WELD WORKFLOW

    weld [options] -- file file ...
    Vigorously fuse files together, without using a cat.

    Fuse two or more files together, absolutely without animal abuse, unlike the
    competitor unix tool cat. Please do not expect deterministic results because
    the printed output is generated by stick electrodes. Please still consider
    the environment before printing the result, with this or with cat.

Explanation of the changes:

Omit that the tool "Robustly" does something. Since somebody could mistook it for
really a stable improvement of cat. Instead, say "Vigorously".

Hints from the beginning that it has something to do with "cat":

    lang=diff
    -Robustly fuse files together.
    +Vigorously fuse files together, withuot using a cat.

Giving this preamble, joke on the fact that we don't use a cat.

> absolutely without animal abuse

Also clarify a bit that the output is a bit random:

> Please do not expect deterministic results

Also give an hint about the output, that is printed (on the screen):

> because the printed output is generated by stick electrodes

...but respect the environment please.

> Please still consider the environment before printing the result ...

Test Plan:
    arc weld --help

Reviewers: O1 Blessed Committers, aklapper

Reviewed By: O1 Blessed Committers, aklapper

Subscribers: aklapper, chris, tobiaswiese, Matthew, Cigaryno

Differential Revision: https://we.phorge.it/D26097