PhutilUTF8StringTruncator: Ignore stop characters not followed by whitespace
Summary:
It's more obvious to readers that some characters were left out when using an ellipsis.
A full stop / dot usually implies the end of a sentence but we do not end a sentence when that full stop is not followed by a whitespace character.

Closes T16293

Test Plan:
Insert the following code in a random place and call it:
```
$string = "uw.controller: Emit 'load' earlier so that mw.UploadWizard#removeError works";
$trunc = id(new PhutilUTF8StringTruncator())
         ->setMaximumGlyphs(70)
         ->truncateString($string);
phlog($trunc);
```

Output before the patch:
`PHLOG: 'uw.controller: Emit 'load' earlier so that mw.'`

Output after the patch:
`PHLOG: 'uw.controller: Emit 'load' earlier so that…'`

Reviewers: O1 Blessed Committers, avivey

Reviewed By: O1 Blessed Committers, avivey

Subscribers: avivey, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16293

Differential Revision: https://we.phorge.it/D26411