PHP-Parser: Name resolve all nodes before traversing
Summary:
Previously, the name resolving was interleaved with the other visitors.
Most of the time this is fine, but in certain cases we seek deep into
the ast, where not all names have been resolved yet, such as trait use
statements. This results in relative names dependent on a use statement,
rather than the fully qualified name that had been expected.

This isn't an issue yet because this code base isn't using namespaces
yet, but will be once it does.

This doesn't seem to slow down the script much if at all.
A full rebuild of the phorge code base is about as quick as before.

Test Plan:
I did not test this directly, but I presume the following:
* Create a phutil library with namespaces and traits
* Use traits with relative names through use statements
* Observe differences in output

Reviewers: O1 Blessed Committers, avivey

Reviewed By: O1 Blessed Committers, avivey

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16289

Differential Revision: https://we.phorge.it/D26456