Allow loading PHAR extensions
Summary:
This appears to be all that's needed to allow it.

For now, not officially supported on Windows, mostly because I don't have a way to test it yet.

This actually works for both Arcanist and Phorge extensions, by just specifying the full path.

On php earlier then 8.0, just refuse everything. I'm pretty sure that at no point do we use web-side inputs as filenames, but this will make it even more sure.

This implementation is a little more limiting in terms of filename allowed for the archive (php allows `*.phar.*`, I think), because we're parsing the file uri ourselves and I think this limit is reasonable.

Refs T16300

Test Plan: Used P56 to make phar file out of R10, added it to `load-libraries` as `phar:///full-path/src`, and it seemed to work.

Reviewers: O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: mainframe98, tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16300

Differential Revision: https://we.phorge.it/D26379